move method to separate package
