+++ satellite/satellitedb/satellitedbtest/utils.go
@@ -1,11 +1,16 @@
 // Copyright (C) 2018 Storj Labs, Inc.
 // See LICENSE for copying information.
 
-package satellitedb
+package satellitedbtest
+
+// This package should be referenced only in test files!
 
 import (
 	"flag"
 	"os"
+	"testing"
+
+	"storj.io/storj/satellite/satellitedb"
 )
 
 const (
@@ -18,14 +23,9 @@ var (
 	testPostgres = flag.String("postgres-test-db", os.Getenv("STORJ_POSTGRES_TEST"), "PostgreSQL test database connection string")
 )
 
-type basicLogger interface {
-	Log(args ...interface{})
-	Fatal(args ...interface{})
-}
-
 // ForEach method will iterate over all supported databases. Will establish
 // connection and will create tables for each DB.
-func ForEach(logger basicLogger, test func(db *DB)) {
+func ForEach(t *testing.T, test func(db *satellitedb.DB)) {
 	for _, dbInfo := range []struct {
 		dbName    string
 		dbURL     string
@@ -34,32 +34,29 @@ func ForEach(logger basicLogger, test func(db *DB)) {
 		{"Sqlite", defaultSqliteConn, ""},
 		{"Postgres", *testPostgres, "Postgres flag missing, example: -postgres-test-db=" + defaultPostgresConn},
 	} {
-		if dbInfo.dbURL == "" {
-			logger.Log("Database", dbInfo.dbName, "connection string not provided.", dbInfo.dbMessage)
-			continue
-		}
-
-		logger.Log("Start testing", dbInfo.dbName)
-
-		db, err := NewDB(dbInfo.dbURL)
-		if err != nil {
-			logger.Fatal(err)
-		}
-
-		defer func() {
-			err := db.Close()
-			if err != nil {
-				logger.Fatal(err)
+		t.Run(dbInfo.dbName, func(t *testing.T) {
+			if dbInfo.dbURL == "" {
+				t.Skipf("Database %s connection string not provided. %s", dbInfo.dbName, dbInfo.dbMessage)
 			}
-		}()
 
-		err = db.CreateTables()
-		if err != nil {
-			logger.Fatal(err)
-		}
+			db, err := satellitedb.NewDB(dbInfo.dbURL)
+			if err != nil {
+				t.Fatal(err)
+			}
 
-		test(db)
+			defer func() {
+				err := db.Close()
+				if err != nil {
+					t.Fatal(err)
+				}
+			}()
 
-		logger.Log("Stop testing", dbInfo.dbName)
+			err = db.CreateTables()
+			if err != nil {
+				t.Fatal(err)
+			}
+
+			test(db)
+		})
 	}
 }
